chore: move function to utils
