+++ app/api/front/conversations/route.ts
@@ -6,16 +6,12 @@ import {
 import { type NextRequest, NextResponse } from "next/server";
 import type { HandoffChatMessage } from "@/types";
 import { HANDOFF_AUTH_HEADER } from "@/app/constants/authentication";
-import type { FrontApplicationClient } from "../client";
 import {
-  convertToFrontMessage,
   createApplicationChannelClient,
-  sendMessageToFront,
+  postMavenMessagesToFront,
 } from "../utils";
 import type { VerifiedUserData } from "@/types";
 import { nanoid } from "nanoid";
-import { JsonFetchError } from "@/lib/jsonFetch";
-import Bottleneck from "bottleneck";
 
 export async function POST(req: NextRequest) {
   return withAppSettings(req, async (request, settings, _orgId, _agentId) => {
@@ -69,63 +65,4 @@ export async function POST(req: NextRequest) {
       },
     );
   });
-}
-
-async function postMavenMessagesToFront({
-  conversationId,
-  client,
-  messages,
-  userInfo,
-}: {
-  conversationId: string;
-  client: FrontApplicationClient;
-  messages: HandoffChatMessage[];
-  userInfo: VerifiedUserData;
-}) {
-  const frontMessages = messages
-    .map((message: any) =>
-      convertToFrontMessage(conversationId, userInfo, message),
-    )
-    .filter((message) => !!message);
-
-  if (!frontMessages.length) {
-    return;
-  }
-
-  const frontLimiter = new Bottleneck({
-    minTime: 200, // 5 requests per second per conversation
-  });
-  frontLimiter.on("failed", async (error, info) => {
-    enum RetryableStatusCodes {
-      TooManyRequests = 429,
-      InternalServerError = 500,
-      NotImplemented = 501,
-      BadGateway = 502,
-      ServiceUnavailable = 503,
-      GatewayTimeout = 504,
-    }
-    const { retryCount } = info;
-    const backoffs = [0.2, 0.4, 0.8, 1, 2];
-    if (
-      error instanceof JsonFetchError &&
-      Object.values(RetryableStatusCodes).includes(error.response.status)
-    ) {
-      const retryAfterSeconds = parseInt(
-        error.response.headers.get("retry-after") ??
-          String(backoffs[retryCount]),
-        10,
-      );
-      return retryAfterSeconds * 1000;
-    }
-    return;
-  });
-
-  for (const message of frontMessages) {
-    try {
-      await frontLimiter.schedule(() => sendMessageToFront(client, message));
-    } catch (error) {
-      console.error(`Failed to deliver message to Front`, error);
-      throw new Error("Failed to deliver message");
-    }
-  }
 }
